src/agency_swarm_Webdesign/requirements.txt
agency-swarm==0.4.4
openai>=1.55.3,<2.0.0
python-dotenv==1.0.1
requests==2.31.0
beautifulsoup4==4.12.2
gradio==3.41.2
langchain==0.3.16
python-dateutil==2.9.0
setuptools>=65.5.1
audioop-lts>=0.2.1 